# [Bugfix] FA2 `inf` workaround with minimum overhead for MLA