@LucasWilkinson 2025-04-02
@DefTruth apologies for the delays! is there a reason you closed this PR? might good to note that here for posterity

in response to the PR though I think we should see if can optimize the kernel with conditional in-place since this seems more robust than doing something outside the kernel