@DefTruth 2025-04-02
doing this outside still cant fully fixed this problem, merge_attn_states kernel is not only be used in MLA but also in flash_attn backend, so, this PR still cant fix all problem. I think you are doing the right things to covert inf to -inf inside the triton kernel. I still try to figure out while these two lines code make performance goes down for R1 AWQ model.



---- Replied Message ----
| From | Lucas ***@***.***> |
| Date | 04/02/2025 22:07 |
| To | ***@***.***> |
| Cc | ***@***.***>***@***.***> |
| Subject | Re: [vllm-project/vllm] [Bugfix] FA2 `inf` workaround with minimum overhead for MLA (PR #15742) |

@DefTruth apologies for the delays! is there a reason you closed this PR? might good to note that here for posterity

in response to the PR though I think we should see if can optimize the kernel with conditional in-place since this seems more robust than doing something outside the kernel

—
Reply to this email directly, view it on GitHub, or unsubscribe.
You are receiving this because you were mentioned.Message ID: ***@***.***>

LucasWilkinson left a comment (vllm-project/vllm#15742)

@DefTruth apologies for the delays! is there a reason you closed this PR? might good to note that here for posterity

in response to the PR though I think we should see if can optimize the kernel with conditional in-place since this seems more robust than doing something outside the kernel

—
Reply to this email directly, view it on GitHub, or unsubscribe.
You are receiving this because you were mentioned.Message ID: ***@***.***>